chore: move ClientIdentifier type to core types
