+++ src/transforms/v2-to-v3/apis/getClientWaiterCallExpression.ts
@@ -1,6 +1,6 @@
 import { CallExpression } from "jscodeshift";
 
-import { ClientIdentifier } from "./getClientIdentifiers";
+import { ClientIdentifier } from "../types";
 
 export const getClientWaiterCallExpression = (
   clientId: ClientIdentifier,
