+++ src/transforms/v2-to-v3/apis/removePromiseCalls.ts
@@ -1,6 +1,6 @@
 import { Collection, Identifier, JSCodeshift } from "jscodeshift";
 
-import { ClientIdentifier } from "./getClientIdentifiers";
+import { ClientIdentifier } from "../types";
 import { removePromiseForCallExpression } from "./removePromiseForCallExpression";
 
 export interface RemovePromiseCallsOptions {
