+++ src/transforms/v2-to-v3/types.ts
@@ -1,3 +1,5 @@
+import { Identifier, ThisExpression } from "jscodeshift";
+
 export type ClientMetadataRecord = Record<string, ClientMetadata>;
 
 export interface ClientMetadata {
@@ -5,3 +7,12 @@ export interface ClientMetadata {
   v3ClientName: string;
   v3ClientPackageName: string;
 }
+
+export interface ThisMemberExpression {
+  type: "MemberExpression";
+  object: ThisExpression;
+  property: Identifier;
+}
+
+export type ClientIdentifier = Identifier | ThisMemberExpression;
+export type ClientIdentifiersRecord = Record<string, ClientIdentifier[]>;
